requirements.txt
qdrant-client==1.7.0
dashscope==1.14.0
numpy==1.24.3
python-dotenv==1.0.0 
packaging
sympy